Remove usages of `simulate_platform(ruby)`
This helper monkeypatches rubygems to return the given platform.
However, "ruby" is not something rubygems ever returns as the current
platform, so this monkeypatch does not represent any real usage. Also,
further refactorings in the resolver were making these specs fail.

Let's use the `force_ruby_platform` setting, which is the closest thing
we have to "current platform is ruby".