Stop cleaning up `$LOAD_PATH` on bundler/setup
If the `$LOAD_PATH` has been modified prior to bundler being loaded,
don't mess with that, just put the paths of the gems resolved by the
Gemfile{,.lock} file in front of it.

This way, if some default gem has been required before bundler, and
rubygems has enhanced the `$LOAD_PATH` to use the latest version in the
system, further requires of that default gem after bundler has been
activated will use the same version. That, unless the gem is a Gemfile
dependency. In that case, we'll get a mismatch error anyways as we do
now.